+++ qiskit/providers/aer/pulse/de_solvers/pulse0_solver_options.py
